+++ ChangeLog.md
@@ -11,6 +11,7 @@
   * Update location event with details of station name and type where available
   * Add events when entering and leaving a station's no-fire zone.  Details about the individual events are in the relevant documentation
   * Add powerplay events.  Details on the new events are in the relevant documentation
+  * Ensure that personality names do not contain illegal file or path characters to avoid issues when saving them
 
 #2.0.13
   * Fix issue where engineer rank journal message without rank would cause a crash